requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
sentence-transformers
torch
transformers
accelerate
faiss-cpu